Going back to having just a models.py file for now
