--- eureka/S5_lightcurve_fitting/models/__init__.py
@@ -1,4 +1,0 @@
-from .Model import Model
-from .TransitModel import TransitModel
-from .PolynomialModel import PolynomialModel
-from .ExpRampModel import ExpRampModel
